requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.4
django-elasticsearch-dsl==7.3
django-filter==23.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
pymongo==4.4.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
